Moving input creation to it’s own file with tests
